Biome Debugging commands added
Manually searching for biomes and mobs is too tedius.
Creating another set of commands to help expidite debugging related to
spawning.
+New commands
/gaia biome list
-Prints a list of all registered biomes in the game
/gaia biome debug <Biome ID>
-Changes a large area of chunks to the specified biome

+added configs for the debugging commands

//Also not quite sure why the ocean mobs won't spawn or register
correctly - if I register them to say a forest they will spawn - so
something odd is going on in the ocean registry - perhaps mojang changed
some rules we need to obey there
//Updating for mesa biomes should be easy though