Change Spriggan translation in ZH_CN
In order to distinguish Dryad and Spriggan in Simplified Chinese version. One of these two creatures' translation must be changed. I prefer changing the translation for Spriggan. Here is my reason:
The wiki of GoG says "The design of the Spriggan is loosely based on the Spriggan from The Elder Scrolls V: Skyrim.". Hence the name of Spriggan in Skyrim is "树精" in Chinese version.
The translation for Dryad is ok in my opinion, because many fantasy animes and games have this creature which commonly translated for "树妖". for example in Warcraft Series (https://warcraft.huijiwiki.com/wiki/%E6%A0%91%E5%A6%96).